+++ test-projects/minecraft/fabric/src/main/java/dev/mavenhaste/compat/FabricCompat.java
@@ -1,0 +1,8 @@
+package dev.mavenhaste.compat;
+
+import net.fabricmc.api.ModInitializer;
+
+public final class FabricCompat implements ModInitializer {
+    @Override
+    public void onInitialize() {}
+}