+++ test-projects/minecraft/forge/src/main/java/dev/mavenhaste/compat/ForgeCompat.java
@@ -1,0 +1,8 @@
+package dev.mavenhaste.compat;
+
+import net.minecraftforge.fml.common.Mod;
+
+@Mod(ForgeCompat.MOD_ID)
+public final class ForgeCompat {
+    public static final String MOD_ID = "maven_haste_forge_compat";
+}